Create a C++ Core Guidelines PDF book
Add the ability to create a PDF book of the C++ Core Guidelines using
pandoc and latex.

Requirements:
* Docker
* Make

How to run:
Run `make` (or `make clean all` if you've made changes) in the linux/
subdirectory. The resulting PDF will be placed in
linux/CppCoreGuidelines.md.

What it does:
1. If it doesn't already exist, Builds a Docker image called
`cppcoreguidelines` that has all of the packages required to build the
book.
2. Builds a PDF of the C++ Core Guidelines in a `cppcoreguidelines`
Docker container using pandoc and latexmk.